reorganize and simplify code
